+++ app/[name]/details/components/additional-info/AdditionalInfo.styles.ts
@@ -1,0 +1,2 @@
+export const AdditionalInfoCont = { fontSize: "1rem", lineHeight: 1.6 };
+export const infoText = { mb: 1 };
