+++ app/[name]/details/components/additional-info/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AdditionalInfo";
