+++ app/[name]/details/components/all-reviews/AllReviews.styles.ts
@@ -1,0 +1,1 @@
+export const title = { mb: 2 };
