+++ app/[name]/details/components/all-reviews/AllReviews.tsx
@@ -1,11 +1,12 @@
 import { Product } from "@/components/types/typeProduct";
 import { Box, Typography } from "@mui/material";
-import Review from "./Review";
+import Review from "../review";
+import { title } from "./AllReviews.styles";
 
 export default function Reviews({ product }: { product: Product }) {
   return (
     <Box sx={{ mt: 3 }}>
-      <Typography variant="h5" component="h2" sx={{ mb: 2 }}>
+      <Typography variant="h5" component="h2" sx={title}>
         Customer Reviews
       </Typography>
