+++ app/[name]/details/components/all-reviews/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./AllReviews";
