+++ app/[name]/details/components/product-info/ProductInfo.styles.ts
@@ -1,0 +1,16 @@
+export const infoBox = {
+  display: "flex",
+  color: "#e43030",
+  alignItems: "center",
+  fontSize: "1.5rem",
+};
+
+export const price = {
+  textDecoration: "line-through",
+  color: "#999",
+  ml: 2,
+};
+
+export const discount = { ml: "auto", fontSize: "1.5rem" };
+
+export const details = { fontSize: "1.2rem", mb: 1 };
