+++ app/[name]/details/components/product-info/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ProductInfo";
