+++ app/[name]/details/components/review/Review.styles.ts
@@ -1,0 +1,11 @@
+export const reviewContainer = {
+  mb: 2,
+  p: 2,
+  border: "1px solid #cecece",
+  borderRadius: "8px",
+  backgroundColor: "#f1f1f1",
+};
+
+export const date = { mt: 1, color: "#888" };
+
+export const text = { mt: 1 };
