+++ app/[name]/details/loading.tsx
@@ -1,4 +1,4 @@
-import Loader from "@/components/Loader";
+import Loader from "@/components/loader";
 
 export default function Loading() {
   return <Loader />;
