+++ app/components/main-page-content/MainPageContent.styles.ts
@@ -1,0 +1,16 @@
+export const styles = {
+  productGrid: {
+    display: "grid",
+    gridTemplateColumns: {
+      xs: "1fr",
+      sm: "1fr 1fr",
+      md: "1fr 1fr",
+      lg: "1fr 1fr",
+    },
+    gap: "20px",
+    padding: "20px",
+    maxWidth: "1200px",
+    margin: "0 auto",
+    gridAutoRows: "1fr",
+  },
+};
