+++ app/components/main-page-content/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./MainPageContent";
