+++ app/components/search-and-sort/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SearchAndSort";
