--- app/globals.css
@@ -1,1 +1,0 @@
-
