+++ app/page.tsx
@@ -1,4 +1,4 @@
-import MainPageContent from "@/app/components/MainPageContent";
+import MainPageContent from "@/app/components/main-page-content";
 import { Box } from "@mui/material";
 
 export default function Home() {
