+++ components/error/Error.styles.ts
@@ -1,0 +1,9 @@
+export const container = {
+  textAlign: "center",
+  marginTop: "50px",
+};
+
+export const text = {
+  fontSize: "2rem",
+  fontStyle: "italic",
+};
