+++ components/header/header.styles.ts
@@ -1,0 +1,11 @@
+export const headerContainer = {
+  backgroundColor: "#443b38",
+};
+
+export const title = {
+  color: "#ecd9d0",
+  flexGrow: 1,
+  textAlign: "center",
+  fontWeight: "bold",
+  padding: "10px",
+};
