+++ components/loader/Loader.styles.ts
@@ -1,0 +1,8 @@
+export const loaderBox = {
+  textAlign: "center",
+  marginTop: "50px",
+};
+
+export const progress = { color: "black" };
+
+export const text = { marginTop: "20px", fontSize: "1.5rem" };
